autodoc refactor: mutualized registration and fields
The main motivation is that Mixin's autodoc_class() was
mostly a duplicate of Model.

Now most of the work is done from the Sphinx extension, unless
there's an autodoc_class() (therefore I expect downstream
libraries that would define autodoc_class() not to be affected).

The behaviour is controled by three new class attributes:
autodoc_anyblok_{field,registration,kwargs}. Also the
defaulting mechanism is simpler (we don't need a wrapper to
present a function as if it were a classmethod for uniformity
any more)