doc: optionally, Bloks readme can be the docstring
This provides a better integration when one also uses
autoclass in the separate documentation: instead of
duplicating part the README.rst in the docstring and including
the README after (or before) the autoclass, one can simply
conflate both.

I'm still keeping the behaviour that 'README.rst' is the default value
in order not to affect downstream code, but I do think this should be the
default.

Now putting lots of Sphinx directives in the docstring would have the
effect to be unreadable outside of Sphinx generation. We'll see
(maybe switching to a more streamlined docstring syntax like Napoleon
or Google style would be nice for AnyBlok itself, haven't explored
them yet, though)