Refactor Model type
The notion of Model Type by plugins is:

* Hard to understand and implement a new version
* Complexify the plugin
* bad naming

This functionnality is replaced by factory class. By default the factory
used is the anyblok.model.factory.ModelFactory. To define a View the dev must use
the anyblok.model.factory.ViewFactory

Define a new Factory and share it become more readable and understanding