Introducing SharedDataTestCase
This new base class is intended for cases where setting up testing
data is long and therefore interesting to be shared among test
cases of a given class.

In principle, this could be merged in BlokTestCase itself: it wouldn't
harm for the shared data to be actually empty.
In practice, we've decided with jssuzanne that we wouldn't like to
impact all existing BlokTestCases right now with potentially subtle
transactional bugs. Hence, we have a new class for now.