Fixed DateTime plugins.
The goal of the datetime plugin, is to update the datetime columns when
and only when the attribute auto_update is True. But the columns from
inherited models have not be seen by the plugins. Now theses columns is
seen and the polymorphics model can use this behaviour on the main
Model.